refactor: replace method syncUrl from ts files to forQueryParam.ts
