+++ src/components/filters/filterByBrand.ts
@@ -1,5 +1,6 @@
 import { IProduct } from '../../types/types';
 import { AllFiltersType } from '../../types/types';
+import { syncURL } from '../forQueryParam/forQueryParam';
 
 export class FilterBrand {
   uploadFilter(): void {
@@ -98,21 +99,6 @@ export class FilterBrand {
   }
   checkFilter(allFiltersOnload?: AllFiltersType): void {
     const filterRow = document.querySelectorAll('.filters__input-row-brand');
-    function transformToURLParams(filters: AllFiltersType) {
-      const query = Object.entries(filters)
-        .map(([key, value]) => {
-          return `${key}=${value}`;
-        })
-        .join('&');
-      return `?${query}`;
-    }
-    function syncURL(filters: AllFiltersType) {
-      const path = document.location.pathname;
-      const query = transformToURLParams(filters);
-      window.history.pushState(filters, '', `${path}${query}`);
-      window.history.pushState(filters, '', `${path}${query}`);
-      history.back();
-    }
     filterRow.forEach((item) => {
       const input = item.children[0] as HTMLInputElement;
       input.addEventListener('change', function () {
