+++ src/components/search/search.ts
@@ -1,5 +1,5 @@
 import { AllFiltersType, IProduct } from '../../types/types';
-import { allFilters } from '../forQueryParam/objOfQueryParam';
+import { allFilters, syncURL } from '../forQueryParam/forQueryParam';
 
 export class Search {
     searchInput: HTMLSelectElement;
@@ -79,20 +79,4 @@ export class Search {
         });
         return searchData;
     }
-}
-
-function transformToURLParams(filters: AllFiltersType) {
-    const query = Object.entries(filters)
-        .map(([key, value]) => {
-        return `${key}=${value}`;
-        })
-        .join('&');
-    return `?${query}`;
-}
-function syncURL(filters: AllFiltersType) {
-    const path = document.location.pathname;
-    const query = transformToURLParams(filters);
-    window.history.pushState(filters, '', `${path}${query}`);
-    window.history.pushState(filters, '', `${path}${query}`);
-    history.back();
 }
