+++ src/components/typeOfView/typeOfView.ts
@@ -1,3 +1,5 @@
+import { allFilters, syncURL } from "../forQueryParam/forQueryParam";
+
 export class typeOfView {
     buttonView: HTMLSelectElement;
     constructor(){
@@ -6,12 +8,33 @@ export class typeOfView {
 
     async addEventListenerButtonView(){
         this.buttonView.addEventListener('click', function(e) {
+            let typeView: string;
             const btn = (document.querySelector('.typeView__icon') as HTMLButtonElement);
             if(btn.innerHTML === '≡'){
                 btn.innerHTML = '⊞';
+                typeView = 'lines';
             }
             else {
                 btn.innerHTML = '≡';
+                typeView = 'blocks'
+            }
+            const searchClear = location.search.split('');
+            searchClear.shift();
+            const queryParamsString = searchClear.join('').toString();
+            let paramsObject;
+            if(queryParamsString !== '') {
+                paramsObject = JSON.parse(
+                    '{"' + decodeURI(queryParamsString).replace(/"/g, '\\"').replace(/&/g, '","').replace(/=/g, '":"') + '"}'
+                );
+            }
+            if(paramsObject !== undefined){
+                if(paramsObject.view !== undefined){
+                    paramsObject.view = typeView;
+                    syncURL(paramsObject);
+                }
+            }
+            else {
+                syncURL(allFilters);
             }
         });
     }
